Implement cursory text to music engine.
Pentatonic and Chromatic scales are provided; notes are being
generated one per letter in a modulo mapping from the alphabet
onto the respective scales, with constant duration and no pauses
between notes. Spacing and punctuation are simply discarded, and
letters are all normalized to lower case.

There is a lot of duplication in this code, to be refactored in
a future commit. Other plans for the future include changing the
sort of music created (specifically to generate chords) and
adding more scales. The synth could use some playing with, too.

The audiolet library is vast and complicated, and I wish it was
not quite so vast and complicated. A library it uses (written by
the same author) has some specialized worker logic with its own
event and timing system, which is apparently necessitated by the
fact that we're dealing with high sample rate audio, which needs
a level of timing accuracy not normally present in javascript.

Unfortunately, there appears to be a bug in that code, which
leads to a buffer-filling function to be called about 12 times a
second forever (until page reload) after even a single note is
played.

I believe that this bug is the source of some popping / cracking
sounds that creep in after a dozen or so notes have been played.
I couldn't find any other good audio synthesizing choices out
there, so the plan is to implement features and just hope it
gets fixed upstream.