Change instances of old name name to new app name.
Soundtrick was a cool idea for a name -- like soundtrack, get it?
It's the soundtrack to the text you're reading! ._. It wasn't very
clear about the app's purpose, and I couldn't ever remember what
the name was when I wasn't reading it on the screen.

So, 'musical words' sounds a bit better.